+++ base/linalg/generic.jl
@@ -739,7 +739,7 @@ of the [`eltype`](@ref) of `A`.
 julia> rank(eye(3))
 3
 
-julia> rank(0 => diagm([1, 0, 2]))
+julia> rank(diagm(0 => [1, 0, 2]))
 2
 
 julia> rank(diagm(0 => [1, 0.001, 2]), 0.1)